pyproject.toml
[project]
name = "news-push"
version = "0.1.0"
description = "FastAPI notifier for daily news images and Chengdu oil prices"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "apscheduler>=3.10.0,<4.0.0",
  "python-docx>=1.1.0,<2.0.0",
  "beautifulsoup4>=4.12.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
news_push = ["data/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
